requirements.txt
psycopg[binary]
requests
python-dotenv
